feat: refactor TemplateEngine into focused single-responsibility classes
BREAKING: None - maintains full backward compatibility

Split the 1645-line TemplateEngine god class into focused services:

- ITemplateLoader/TemplateLoader: Load templates from resources
- ITemplateParser/TemplateParser: Parse template syntax and normalize names
- ITemplateRenderer/TemplateRenderer: Render templates with variable substitution
- ITemplateValidator/TemplateValidator: Validate templates and project options
- ITemplateCache/TemplateCache: Cache frequently used templates
- TemplateServices: DI container for service management

Benefits:
✅ Single Responsibility Principle
✅ Better testability with focused classes
✅ Improved maintainability and extensibility
✅ Built-in validation with detailed error messages
✅ Template caching for better performance
✅ Full backward compatibility - no API changes

Files:
- Backed up original TemplateEngine.cs
- Added Microsoft.Extensions.DependencyInjection dependency
- Created comprehensive interfaces and implementations
- Added validation tests and documentation

Resolves: #1181
